перенес операции с фильмами и юзерами в storage сервисы
