+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -5,6 +5,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.UserService;
+import ru.yandex.practicum.filmorate.storage.user.InMemoryUserStorage;
 
 import java.util.Collection;
 
@@ -13,20 +14,20 @@ import java.util.Collection;
 @RequiredArgsConstructor
 public class UserController {
 
-    private final UserService userService;
+    private final InMemoryUserStorage userStorage;
 
     @GetMapping
     public Collection<User> getUsers() {
-        return userService.getUsers();
+        return userStorage.getUsers();
     }
 
     @PostMapping
     public User createUser(@Valid @RequestBody User user) {
-        return userService.addUser(user);
+        return userStorage.addUser(user);
     }
 
     @PutMapping
     public User updateUser(@Valid @RequestBody User newUser) {
-        return userService.updateUser(newUser);
+        return userStorage.updateUser(newUser);
     }
 }
