+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -14,63 +14,5 @@ import java.util.Optional;
 @Service
 @Slf4j
 public class UserService {
-    private final Map<Long, User> users = new HashMap<>();
 
-    public Collection<User> getUsers() {
-        return users.values();
-    }
-
-    public Map<Long, User> getUsersMap() {
-        return users;
-    }
-
-    public User addUser(User user) {
-        user.setId(getNextId());
-        if (user.getName() == null || user.getName().isBlank()) {
-            user.setName(user.getLogin());
-            log.info("В запросе создания пользователя передан пустой name. Для поля name использован логин {}",
-                    user.getLogin());
-        }
-        users.put(user.getId(), user);
-        log.info("Добавлен пользователь: {}", user);
-        return user;
-    }
-
-    public User updateUser(User newUser) {
-        if (newUser.getId() == null) {
-            log.warn("В запросе на обновление юзера не передан id");
-            throw new ValidationException("Id не может быть пустым");
-        }
-        if (users.containsKey(newUser.getId())) {
-            User oldUser = users.get(newUser.getId());
-            if (newUser.getName() == null || newUser.getName().isBlank()) {
-                oldUser.setName(newUser.getLogin());
-                log.info("В запросе обновления пользователя передан пустой name. Для поля name использован логин {}",
-                        newUser.getLogin());
-            } else {
-                Optional.of(newUser.getName()).ifPresent(oldUser::setName);
-            }
-            Optional.ofNullable(newUser.getEmail()).ifPresent(oldUser::setEmail);
-            Optional.ofNullable(newUser.getBirthday()).ifPresent(oldUser::setBirthday);
-            Optional.ofNullable(newUser.getLogin()).ifPresent(oldUser::setLogin);
-            log.info("Обновлен юзер: {}", oldUser);
-            return oldUser;
-        } else {
-            log.warn("В запросе на обновление пользователя передан неизвестный id - {}", newUser.getId());
-            throw new NotFoundException("Пользователь с id " + newUser.getId() + " не найден");
-        }
-    }
-
-    public void clearData() {
-        users.clear();
-    }
-
-    private long getNextId() {
-        long currentMaxId = users.keySet()
-                .stream()
-                .mapToLong(id -> id)
-                .max()
-                .orElse(0);
-        return ++currentMaxId;
-    }
 }
