+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -1,0 +1,16 @@
+package ru.yandex.practicum.filmorate.storage.user;
+
+import ru.yandex.practicum.filmorate.model.User;
+
+import java.util.Collection;
+
+public interface UserStorage {
+
+    User addUser(User user);
+
+    User updateUser(User user);
+
+    Collection<User> getUsers();
+
+    User getUser(Long id);
+}
